Setup publish to dist-branch
This will push output of `npm pack` to the `dist` folder, on every push to `main`. This will allow us to use the latest version even before cutting a new release. This is needed because we have a build step, so we cannot just point our `package.json` to the github repo with the source files directly.